Add support for envoy 1.12+ transport_socket tls context
The 'tls_context' key in cluster configs is deprecated since envoy 1.12 and
is removed since 1.14.

This patch detects envoy version and switch tls context setup to the new
transport socket for envoy versions >=1.12

The docker image is also updated to embed python requests, and to fix a
grpcio compilation issue with newer alpine.